Move response.ts to "lib" folder
